Facet source-version: demote the racy pin to a deterministic contract test
The pin asserted a running facet NEVER picks up a source commit; CI proved
that behavior is a race (both outcomes on one deployment minutes apart), so
the failing() wrapper flip-flopped. Keep the deterministic contract as a
plain test (stable key while unchanged; fresh facet after a kill serves the
committed source), delete the blind-spot companion pin, move the racy
pickup bug to tasks/facet-commit-pickup-race.md, and teach the failing()
docs that only deterministic bugs can be pinned.
